+++ drafter/__init__.py
@@ -10,7 +10,7 @@ import drafter.hacks
 # Provide default route
 route('index')(default_index)
 
-__version__ = '1.4.2'
+__version__ = '1.5.0'
 
 if __name__ == '__main__':
     print("This package is meant to be imported, not run as a script. For now, at least.")