+++ examples/file_upload_testing.py
@@ -1,12 +1,8 @@
+from bakery import assert_equal
 from drafter import *
 from dataclasses import dataclass
 from PIL import Image as PILImage
-import io
-import base64
 
-RAW_SIMPLE_IMAGE = '''R0lGODlhDwAPAKECAAAAzMzM/////wAAACwAAAAADwAPAAACIISPeQHsrZ5ModrLl
-N48CXF8m2iQ3YmmKqVlRtW4MLwWACH+H09wdGltaXplZCBieSBVbGVhZCBTbWFydFNhdmVyIQAAOw=='''
-SIMPLE_IMAGE = PILImage.open(io.BytesIO(base64.b64decode(RAW_SIMPLE_IMAGE)))
 
 @dataclass
 class State:
@@ -60,5 +56,31 @@ def download_page(state: State) -> Page:
     ])
 
 
-start_server(State("NO TEXT YET", b"NO BITS YET".decode('utf-8'), SIMPLE_IMAGE),
-             src_image_folder="images")
+inital_state = State("", "", PILImage.new("RGB", (1, 1)))
+assert_equal(update_data(inital_state, "Hello, world!", b"01010101", PILImage.new("RGB", (100, 100))),
+             Page(State("Hello, world!", "b'01010101'", PILImage.new("RGB", (100, 100))), [
+                 "You can upload three things here:",
+                 HorizontalRule(),
+                 Div(
+                     Row("Text File:", FileUpload("new_text")),
+                     "Here's the text you uploaded:",
+                     Div("Hello, world!"),
+                     style_margin="4px", style_border="1px solid black"),
+                 HorizontalRule(),
+                 Div(
+                     Row("Bits File:", FileUpload("new_bits")),
+                     "Here are the bits you uploaded:",
+                     PreformattedText("b'01010101'"),
+                     style_margin="4px", style_border="1px solid black"),
+                 HorizontalRule(),
+                 Div(
+                     Row("Image File:", FileUpload("new_image", accept="image/*")),
+                     "Here's the image you uploaded:",
+                     Image(PILImage.new("RGB", (100, 100))),
+                     style_margin="4px", style_border="1px solid black"),
+                 HorizontalRule(),
+                 Button("Update to uploaded files", "update_data"),
+                 Button("Go to download page", "download_page")
+             ]))
+
+# assert_equal(update_data(inital_state,